apps/api/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.30.*
SQLAlchemy==2.0.*
alembic==1.13.*
psycopg2-binary==2.9.*
pydantic==2.8.*
python-multipart==0.0.9
passlib[bcrypt]==1.7.*
pyjwt==2.9.*
python-slugify==8.*
email-validator==2.1.*
python-dotenv==1.0.*
apscheduler==3.10.4
pytz==2024.1
openpyxl==3.1.*
